znaphodl: replace `md5` with `sha256sum`
The `znaphodl` executable has always used `md5` to avoid lockfile
collisions. That was fine because such collisions wouldn’t have been
a security issue.

With the MD5 algorithm falling out of favor though, the binary is no
longer included in some Linux distributions, such as Arch Linux or
the AUR.
Because we already depend on coreutils and the `md5` utility , it makes
sense to use sha256sum for the lockfile instead of md5.